requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
discord.py-rw==1.0.0a0.post2
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
websockets==6.0
yarl==1.3.0